pyproject.toml
[project]
name = "miact"
version = "0.1.0"
description = "Multi-source Information Aggregator & Comparison Tool"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "psycopg2-binary",
    "python-dotenv",
    "beautifulsoup4",
    "requests",
    "newspaper3k",
    "spacy",
    "vaderSentiment",
    "trafilatura",
    "playwright",
    "duckduckgo_search",
    "pandas",
    "lxml",
    "lxml_html_clean",
    "transformers",
    "torch",
    "onnxruntime",
    "optimum[onnxruntime]",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }

[tool.uv]
dev-dependencies = [
    "pytest",
]
